Create TR_update_avg_rating_TriggerTest.sql
Test to ensure that the trigger is working. It first deletes all the other rows from the database, then proceeds to make a single movie, add a user, and make the user rate the movie, then the average rating for the movie goes up according to the user's new rating.